requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
attrs==24.2.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cloudscraper==1.2.71
curl_cffi==0.7.3
frozenlist==1.5.0
g4f==0.3.3.3
idna==3.10
MouseInfo==0.1.3
mss==10.0.0
multidict==6.1.0
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
propcache==0.2.0
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.21.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.2.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python3-xlib==0.15
pytweening==1.2.0
requests==2.32.3
requests-toolbelt==1.0.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.17.1
python-dotenv==1.2.2
pytest==7.3.2
